Refactor tagging imported licences
Tidies up the `LicenceImporter` class and split outs the tagging
functionality into its own class. The aim of this tidy up is to make the
`LicenceImporter` class easier to follow.

A factory needed to be created to test that the licence transaction was
being tagged with the correct locations and industries.
The existing `licence_transaction` factory could not be used as this
factory mocks the payload to publishing-api and returns a Hash object
rather than a LicenceTransaction object.

It looks as though all of the other factories follow the same pattern.
Even though these factories have a "parent" of document, it is not the
document model, but the common parts of the publishing-api payload.

Some consideration was given to renaming these factories to "*_payload"
i.e. "licence_transaction_payload", however considering that this import
code is temporary and will be gone within months, and considering the
sheer number of affected factories, it seemed out of scope for this
change. A new `licence_transaction_model` factory was created instead
for use in this test that can be easily removed with the rest of the
licence importer code.

The facet tagging tests have been removed from the licence importer
tests as those tests test whether the facets are presented to
publishing-api correctly and that is covered by the DocumentPresenter
tests that cover the inclusion of format specific fields (facets).