Remove call method
The update schema and update licence transactions methods cannot be called together.
The update schema method needs to be called locally, and the new schema merged into main.

Then the update licence transactions needs to be called in a environment to update and
republish the licence transactions.

The two methods still need to share methods to know which licences need to be updated
and how.

This process does feel quite brittle though.