Republish the licence transaction
Republish the licence transaction after retagging the industry.

It turns out that the DocumentPublisher does another call to get content in publishing-api
after the document has been published so that it can send an Email alert notifying users
of a change. Therefore the test has been updated to reflect that get content is called twice.

It's strange that email alert api is being called directly here, and not using the flow in
publishing-api. Investigating the email alert flow is out of scope for this change.

An email is only sent if the update type is "major". The state has been left as "draft" as these are
new documents and therefore don't have subscribers.
Licence transactions have not been migrated yet.