Update LicenceImporter to use mapping file
We're now using a mapping file for the tagging of industries / locations
to licences, rather than licencefinders version.

The industries have changed significantly enough that using
licencefinders industries mapping was going to be cumberstone. Having a
complete mapping file is also better as the mapping file includes the
tagging for all the licences we're importing and makes it easier /
possible to update the tagging.

Some rows use the same base_path and so have been grouped and their
industries / locations combinded due to the original spreadsheet [1]
mapping different licences to the same link.

[1]: https://docs.google.com/spreadsheets/d/1ZfGKnVKmfB5yGBWMx3PvwKq8KEj6Lt1XzcWXCJWBMf0/edit#gid=1760190301