Strips licence identifier if continuation link
If a continuation link is present then the external website template is
rendered which doesn't require the licence identifier, as the identifier
is used for non-external link journeys (postcode lookup).

This work is a prerequisite for adding validation to the bespoke fields
for licences (continuation_link, will_continue_on and
licence_identifier).

Trello:
https://trello.com/c/p1DbDyxZ/2021-strip-licence-identifier-from-imported-licences-if-continuation-link-present