Add custom error locales for basic licence fields
Allows a completely custom error message to be used in for the listed
fields for certain validation types. Currently, the implementation is a
bit clumsy but this is a consequence of custom error messages.

We'll add more fields to this list bespoke to licences and there will be
a follow up piece of work to convert all the error messages over to the
new and recommended design system style.