Adds uniqueness validator for licence identifier
We expect each licence to either have a blank licence identifier (if
linked to an external website) or a unique identifier if managed by a
local authority. The validator pulls in all the licences from Publishing
API and checks whether the identifier, if present, already exists and
fails if it does.