Switch to using FinderLoader to load finder schema file when unpublishing finder
To keep consistency with publishing and republishing, we use the FinderLoader.
However, the FinderLoader looks up finder schema via filename, not slug. This functionality needs to be updated in the FinderLoader at a later point to avoid any trip-ups as the filename is not the same as the slug base path in quite a few of the specialist publisher finders.