Retag BEIS finder as DBT
The [Finance and support for your business][1] finder was dual-tagged as
Department for Business, Energy & Industrial Strategy (BEIS) and
Department for Business and Trade (DBT).

However BEIS is now closed and has been replaced by DBT, so we need to
retag this finder accordingly.

This also removes an unrelated organisation which had seeminly
accidentally been granted editing permissions to these documents.

After this change, DBT will be the only department with permission to
edit these documents, and the only one tagged on the frontend too.

[1]: https://www.gov.uk/business-finance-support