Add organisation filter option
We can use `link_organisations` which tells Publishing API to fetch all
documents that have are tagged to the passed in organisation content_id.

`organisations` include both the primary publishing org and the other
tagged organisations.

We only want to filter if the document type `has_organisations?` == true
which denotes the document type's organisations can be chosen by the
user (3 in total). Normal specialist documents don't have this option so
allowing them to filter would be confusing if there is just one or two.

We will filter (on load)  by the user's current signon organisation,
unless another is selected.