Adds organisation filter and org helper method
This is using the select2 CSS / JS which is what we currently use for
organisation select fields, e.g [1]. This seems to fit the design of the
index page well and offers a searchable select box with match
highlighting.

We're also providing an 'All organisations' option which will allow
users to see all documents from all organisations for an eligible type.

Lastly, we're now using instance vars instead of passing `params` around in
the view as this was a little unusual.

[1]: https://github.com/alphagov/specialist-publisher/blob/c98278cafa970ba71fe8f339e1aa8e20708d1824/app/views/metadata_fields/_licences.html.erb#L48