Show link/identifier errors at the related fields
We now display the link/identifier exists (or blank) errors at the
field group they apply to. Previously we only displayed these errors at
the main error summary at the top of the page.

Also moves the 'How users apply' section to the bottom of page which
makes the grouping more obvious.