Add format validation for licence identifier
We've previously been informed of a rendering issue that was caused by
the publisher inputting random text into the licence identifier field.
Previously there was only a uniqueness validator for the field which
accepted any format, now we're validating that the format is correct.

The accepted format is XXX-Y-B or XXXX-Y-B where:
- X is a formality code (3 or 4 digits)
- Y is a country code (1 to 7, 1 digit)
- B is a number in sequence (if more than 1 'type' [quite rare], 1 digit)

Trello:
https://trello.com/c/uFIvraIX/2288-validate-licence-identifer-format