Manaully require files in for lib
The config.add_autoload_paths_to_load_path option is moving from true
to false in 7.1.

This means that simply calling require '{filename}' in spec/lib files won't
work as it searches for the file in the ../lib dir which isn't how
our dir's are structured.

We have a few options:

1. we could set 'config.add_autoload_paths_to_load_path option' to true
2. we could manually update files in the lib folder to require the correct
file
3. we can update the spec helper to require all files in app/lib

It feels to me like if it's not a massive hassle we should probably retain
Rails default config so 2 or 3 are better options.

While there are not many require paths that need to be updated (3 in total),
i suspect it'll be far easier for developers to not have to manaully require files
each time a spec/lib/*_test.rb file is added.