Allow deletion of attachments where the asset does not exist in asset manager
An attachment which does not have a corresponding asset in Asset Manager is effectively broken. Prior to this commit, it wasn't possible for users to delete these broken attachments.

Being unable to delete the broken attachments prevents users from discarding any draft documents associated with them.

To resolve this, we return a boolean "true" from the Attachment#destroy method when the asset cannot be found, as to all intents and purposes the asset is deleted anyway.

It is not clear how the asset has been removed from Asset Manager whilst the attachment is still present.