Remove internal note feature
The functionality for adding an internal note was developed for the UK Market Conformity
Assessment Bodies document type. The finder was unpublished alongside its documents,
and in order to support that process, the internal note served the purpose of documenting
the reasons why a specific document was unpublished.

We are now removing the finder code in order to prevent accidentally republishing it.
Since there are no current usages of the feature as it has not been used for other
finders and it is not a development priority or business request, we are removing all
associated internal note implementation.