WIP: move Specialist Document classes into subfolder
This feels neater than keeping all models at the top level, and
clearly distinguishes these classes from others in the same
directory (namely Organisation, Attachment, AttachmentCollection,
User, and perhaps confusingly, the Document parent class).

But we have to update a LOT of code and workflows to get this to
work - so probably not worth the risk.