Updating docs on editing field values.
The docuemntation seems to only work for adding new values to the facet rather than amending existing ones, which I assume will also entail editing and republishing all existing documents.
in adding new values to the facet, we should only need to update the finder, not the search API schema. This has been tested in integration.