Add "View on website (opens in new tab)" link to index pages
Screenshot:

![screenshot of AAIB report index page](https://github.com/user-attachments/assets/f64ef2b0-1ee2-4fc7-ae92-72f94e78da50)

I did experiment with aligning the link with the `<h1>` as per
the design, but our version of Bootstrap doesn't allow for that
very nicely. I could wrap the `<h1>` and the link in `col-md-9` and
`col-md-3` divs respectively, but the link would be vertically
aligned to the top rather than the bottom, and there didn't appear
to be a way of changing that other than to hack in some flexbox
styles. It also led to questions around how wide each column should
be (different title lengths, different screen sizes, will lead to
different experiences), so stacking the elements seemed the pragmatic
option.