WIP: add pseudo self serve
Adds an `/admin/:document_type_slug` path, which reads the config
from the associated schema JSON file, and displays it in basic
form inputs.

Clicking "Generate schema" generates a JSON string which the user
can copy and paste. In theory, this makes requesting edits to
finders easier, as users can see the existing configuration, make
tweaks, and then give us the tweaked configuration in a ready to
use format. It also paves the way for self-serve in future, where
we may want to automatically republish finders after edits.