Add disable_email_alert flag on the document object
This flag is used to override whether or not emails should be sent for a document upon publishing. The plan right now is to use this flag in bulk publishing rake tasks so we can silence alerts when publishing in large amounts, but in the future could potentially be used within a self-serve UI. However, we need to note that the UI option might be complex as data for specialist publisher is persisted in Publishing API and not internally. Therefore the UI solution needs to consider the fact that the flag is not yet persisted within Publishing API.