Set up metadata summary card for confirm page
If a `previous_schema` is provided to the metadata summary, only show the values that have changed. This will be used for a confirmation page for change requests.

Display related links in metadata summary card; rather than just displaying yes/no, show a list of related links.

Remove the email alerts section as we are moving it to its own form. https://trello.com/c/JoK7tUbQ/3036-release-the-edit-finder-forms-feature-%F0%9F%9A%80.