Set up metadata summary card for confirm page
If a `previous_schema` is provided to the metadata summary, only show the values that have changed. This will be used for a confirmation page for change requests.

Remove the email alerts section as we are moving it to its own form.

Display related links in metadata summary card; rather than just displaying yes/no, show a list of related links.