Add page to edit facets
Introduce route and form to modify the facets for a finder.

We did attempt to have a nice "conditional reveal" on the filter
options for multi/single select, but that doesn't play nicely
with the "Add another" component, so we had to revert it. The
code where we tried it out is in commit:
fcd2e8f45161af9e21618e14f169c906c283b9d8

The other area where the form differs from the design is in the
handling of the 'keys'/'values' (as opposed to 'labels') of the
filter options. The result is the least worst solution to the
problem of: "how do we retain the original filter option key for
_existing_ filter options?".

The design called for a simple textarea comma-separated list of
filter option _labels_ (i.e. "Commercial - fixed wing"), with
nowhere to specify the _key_ / _value_ ("commercial-fixed-wing").
Whilst in that example we can easily convert the label into the
key, there are plenty of examples of existing filter options where
the key is markedly different from the label, and thus submitting
the form would be "lossy" and lead to a larger and unwanted diff.
(Moreover it could lead to real problems if a dev were to go
ahead and merge the diff).

I did explore exposing the key / value pairs in dedicated input
fields, and in theory could then turn the key into an `input
type=hidden` so that it's not exposed to the user. This was
explored in e8810b6a2dc77dbf685aff65bf92f7b11fe6bb36 but the
component doesn't currently allow nesting, so this isn't a
viable way forward for now.

What I've arrived at is I think a good compromise: it is the
label that appears first (followed by the key), and there is no
expectation for the user to edit the key nor to provide a key for
any new filter options they add. The controller automatically
defaults to a kebab-cased key derived from the label, if one is
not provided up front.

(Side note: I couldn't find a Rails native kebabcase method, to
my surprise, so took the regex from
https://jonathanyeong.com/changing-to-kebabcase/).