Point developer to the schema that needs updating
The Zendesk integration currently just outputs the schema contents
that need copying and pasting. But the developer still needs to
track down the file in specialist-publisher, which is an
unnecessarily manual step - let's give them the link directly
to the file edit screen, ready to copy and paste the schema.

Trello: https://trello.com/c/YdGWsy0I/3035-integrate-the-specialist-finder-edit-forms-with-support-api