Create a model class for finder schemas
Without this model, we don't have a central place to keep logic relating to finder schemas, resulting in it leaking out all over the codebase (see the admin controller).

The model can be instantiated from the JSON finder schema, but we can also add logic to build it from the admin metadata form.