Remove signup title schema attribute
The signup title schema attribute was only being used for one finder, and the fallback value was the same as the value provided anyway.

We are therefore removing the signup title schema attribute so that we do not have to support it within the finder configuration editing form that we are building.