Handle email alert form inputs
This was a bit fiddly. The conditional field feature in the GOV.UK radio button component does not prevent the inclusion of the field values in the form submission, because it only hides the conditional DOM using CSS.

We therefore need to filter out any params for fields that should have been excluded based on the email alerts condition in the controller.

We also have to reset the values from the original schema, as we want the values from the form to overwrite the schema values if they change. We use the update method callback to do this.

I took the opportunity to update the signup_copy using an update callback as well, which will ensure we don't get any strange behaviour if the document noun and signup copy are assigned in an order we're not expecting.