Ensure finder schema is output to Zendesk correctly
Now that the finder schema is a model and not a hash, we need to call `to_json` on it to render it inside the proposed schema hidden input value.

The `to_json` method is defined in the ActiveModel::Serializers::JSON module. To use it, I had to redefine the finder schema attributes using ActiveModel::Attributes.

The attributes usage is a little bit peculiar in that you can't write to them from inside the model, you have to use `#assign_attributes` instead. I tried all of `attribute =`, `self.attribute =`, `@attributes[:attribute] =` and `self[:attribute] =` without success.