Introduce form object for handling email alert finder schema values
We need the subscription title list prefix attribute to be provided for both the filtered content email signup and the all content email signup.

However, having duplicate fields in the HTML form, even when one was hidden via the radio button Publishing component conditional behaviour, would result in both input values being submitted.

I have therefore reluctantly chosen to introduce an extra layer of indirection to manage this, as the input form fields will need to have prefixed names. The email alert class handles mapping the form parameter values to the finder schema keys.

The before update callback on the finder schema which reset the email alert attributes is now redundant, so that has been removed.

The input for the subscription title list prefix will be added in the next commit.