Add new 'specialist_publisher_properties' hash to finder schema
This will encapsulate properties we don't intend to surface to
Publishing API nor to Search API - they're used only in the
Specialist Publisher interface. This hash presents an opportunity
to DRY up our views and validation logic, by encapsulating as
much data as possible in the finder schema. This should mean that
minimal / zero edits will be required outside of the schema (in
this repo at least) when editing fields.