Add 'specialist_publisher_properties' to the remaining finders
Cross-referencing with the schema in Publishing API:
https://github.com/alphagov/publishing-api/blob/8d04f4b8629c9834e0d6cb2bcdc8c86ea77372eb/content_schemas/formats/shared/definitions/_specialist_document.jsonnet

For every JSON file in lib/documents/schemas, for every "facets" hash,
where there is a hash with an "allowed_values" property, we've added
an additional property below "allowed_values":

```
"specialist_publisher_properties": {
  "select": "multiple"
}
```

It will be `"select": "multiple"` or `"select": "one"`, depending
on the corresponding config in publishing-api. If the
corresponding config is `type: "array"`, then select multiple, but
if it's just `type: "string"`, then select one.

This follows the example in aaib_reports.json where I applied this
in an earlier commit.

Note that at the time of this commit, not every finder is using
the new specialist publisher properties config, but there's no
harm in getting the properties in ready, and they do help to
document the intended behaviour for each field.