Have simple "specialist document" types use the shared template
These are the specialist document types that we were able to have
pointing to the new generic shared template, with little or no
change in how the form is rendered. It also meant we were able to
delete a "dropdown.html.erb" partial which is no longer being
referenced anywhere.

The handful of views that we haven't touched yet are a bit more
bespoke and will be tackled in subsequent commits.