Update tests to reflect changes in the forms
Switching to the generic form template does lead to some small
differences in the rendered form - nothing big enough to worry
about fixing on a per-finder basis. It's still clear to the
publisher what each form field should do, and it's easy for us
to change (via the finder schema) should we need to. What we have
now is more consistency between what's in the schema, and what's
in the form the publisher sees.