Validate that every facet is actually used in Publishing API
I noticed that there were several facet fields defined in the
tribunal related finders, which aren't actually displayed in
the publishing UI, nor are they referenced in Publishing API.
(Presumably they were superseded by something else but never
removed).

This commit adds validation that checks for that case,
cross-referencing every facet specified in the finder schema
with a corresponding property on
https://github.com/alphagov/publishing-api/blob/main/content_schemas/formats/shared/definitions/_specialist_document.jsonnet

The closest equivalent we have to this kind of validation so
far is the 'find_content_item_presenter_spec.rb'. I did
consider adding this test into that, but we're really just
sense-checking that the finder schemas are valid, not testing
the `FinderContentItemPresenter`, so it didn't feel appropriate
lumping it into that.