Add special case for flood and coastal finder, and remove config
The _only_ finder that doesn't seem to follow the
`{finder_name}_metadata` pattern in Publishing API is the
flood and coastal erosion risk finder. I've added a specific
case to handle that, and a TODO for us to tidy it up later.

There was one property defined on the flood and coastal finder,
which is not referenced in Publishing API, so can be safely
removed.