Add 'hidden_indexable_content' field to finder schemas
This was omitted from the schemas, instead being declared only
in the FORMAT_SPECIFIC_FIELDS const in the relevant models (e.g.
https://github.com/alphagov/specialist-publisher/blob/ec317d198ea68cc779cca7b0b85b2a3a05b13788/app/models/asylum_support_decision.rb#L10).
This was to stop the facet from being included in the finder
content item, but we can now achieve the same thing by using
the "specialist_publisher_properties" configuration.

Moving the field definitions into the finder schemas means we'll
be able to delete all of the views for these and use the shared
view instead (treating it as a common text input like any other
field).

NB, the field is a hard-coded concept in several places:

1. Publishing API: https://github.com/alphagov/publishing-api/blob/303baf6b3f7b083722043054e6af94b0c3b50323/content_schemas/formats/shared/definitions/_specialist_document.jsonnet#L343-L345
2. Search API: https://github.com/alphagov/search-api/blob/7b823c6e369fd27cd5f8f9d80d9095542b420c44/lib/govuk_index/presenters/indexable_content_presenter.rb#L34
3. Specialist Publisher: https://github.com/alphagov/specialist-publisher/blob/58b5b5ea0d5fb34c3de56f43a1264b7effa90012/app/presenters/email_alert_presenter.rb#L36