All finders with hidden_indexable_content use the shared view
See previous commits. Now that the hidden_indexable_content field
has been added to the finder schemas themselves, there's nothing
unqiue about the views for these finder publishing UIs, so they
can start using the shared view.

Note that hidden_indexable_content fields tend to have a lot of
text input (see https://www.gov.uk/api/content/asylum-support-tribunal-decisions/aw-and-am-and-bg-v-secretary-of-state-home-office-as-slash-23-slash-11-slash-45804-and-as-slash-23-slash-11-slash-45824-and-as-slash-23-slash-11-slash-45796
for an example). So I've swapped out the input type for text,
from `input type="text"` to an adjustable `textarea`. This should
have a negligible impact elsewhere - publishers can still just put
a little bit of text in rather than a lot - but this also enables
publishers to continue to use hidden_indexable_content without
restriction.

Moreover, these forms aren't yet on the Design System - so let's
not worry too much about how they look. We can revisit this
properly when we migrate to the new design system.