Added Facet model to keep logic outside of AdminController
Following the same pattern as the EmailAlert model introduced in
ae63aa74d931388dd8b874f4461d18c977ff2c61, we've added a Facet
model which provides a translation layer between the form inputs
we expect to add in subsequent commits, and the eventual
representation of the facet in the finder schema JSON.

This logic started off life inline in the AdminController, which
added bloat and complexity to the controller and would have been
extremely difficult to test. The logic is now instead encapsulated
in "Facet" and has an accompanying test suite, so we can be
confident that form inputs are being correctly converted to their
finder schema JSON form.

See next commit for where this class is actually used.