Remove unused 'subscription_list_title_prefix.many' property
This was added in https://github.com/alphagov/specialist-publisher/pull/986
in January 2017. It was intended to be used when the generated
email subscriber list title would otherwise be longer than 255
characters, which was the maximum allowed by GovDelivery at the
time. Instead of a title with the structure "{document type} with
the following categories: a, b, c, d, e, f...", the 'many' option
generated a title of the structure "{document type}: n", where 'n'
is the number of categories.

This logic never ended up getting used! The author of the commit
comments in their PR:
> following this PR I will raise another PR for finder-frontend
> that splits the logic for title and short_name to be separate so
> that we can continue to set title as we do now and use new logic
> for short_name that limits it to 255 characters.

That corresponding PR appears to be https://github.com/alphagov/finder-frontend/pull/284,
which was closed without merging, with the comment:
> I'm closing this because it's becoming too complicated and the
> email team are looking at changing this all anyway.

Indeed, finder-frontend only seems to make reference to 'singular'
or 'plural' in its email subscriber list title builder:
https://github.com/alphagov/finder-frontend/blob/484396b660715811500af3e3b226ecf0dd69c675/app/lib/email_alert_title_builder.rb#L30-L32

It's therefore safe to remove these unused properties.