Update the finder schemas with the data structures we want
There was a lot of room for improvement, DRYing up of data,
and opportunities to make data more consistent between the
finder itself and the email signup page. This commit:

- Removes `email_filter_facets`: this is now derived from the
  `facets` further down in the schema.
- It's not as simple as that, of course. There are
  inconsistencies between `email_filter_facets` and the
  corresponding `facets`, which I've had to add some overrides
  for by way of a `email_filter_options` hash.

The `email_filter_options` hash has these properties:

- `email_filter_by`, which has been moved down from the top
  level hash.
- `downcase_email_alert_topic_names`: when the generated
  `email_filter_facets.facet_choices` array has a `topic_name`
  that differs from the `radio_button_name` ONLY in its
  casing. A lot of finders have the two identical, but a lot
  of finders seem to have the `topic_name` fully downcased,
  so I've added a boolean option for that here.
- `email_alert_topic_name_overrides`: in rarer cases, a
  facet's `topic_name` is neither the same as the radio button
  name NOR its downcased version - it's slightly different and
  bespoke. For these cases, I've defined a hash which should
  be used to overwrite the topic names for these specific
  facet options.
- `pre_checked_email_alert_checkboxes`: an array of facet keys
  for the facets which should have `prechecked: true` set on
  them. This is only used for the medical safety alerts finder.
- `all_selected_facets_except_for`. Whereas most finders that
  set an `email_filter_by: "all_selected_facets"` mean this to
  grab ALL facets that have `filterable: true` and an
  `allowed_values` array, there is one exception, in the export
  health certificates finder, in which there is a "certificate
  status" facet that follows that pattern, but which was
  not included in the original "email_filter_facets". We could
  investigate removing this hack and just having that signup
  page have the extra facet options, but I wanted to minimise
  any functional changes introduced in this refactor.

NB, `required: true` seems to always be set if `email_filter_by`
is anything other than `all_selected_facets`, and if that
property is something else, then `required: true` NEVER seems
to be set. So we can definitely automate that in the next
commit.

Also worth noting, I started off by moving all these override
parameters into the relevant facet (we have a
"specialist_publisher_properties" hash that serves as a nice home
for that kind of metadata). However, on second thoughts, I didn't
want to complicate the existing "facets" hash any further, as we'd
also have to update the "request edits to filters and options"
form. It's not necessary to do that - the email alert options are
fairly self contained, they just also _refer_ to facets. So I opted
for a new "email_filter_options" hash which can provide overrides
where referring to the raw facets is not enough.