Update test to compare generated content items with snapshots
Rather than creating the snapshots, we now want to have the test
comparing the latest presenter output with the snapshots. (With
the exception of the public_updated_at property, which is always
going to differ).

With this test passing, we're in a good place to start refactoring
the data structure.