Get content item before/after tests passing with new config
As expected, most of the config could be generated / derived from
elsewhere in the finder schema, with some overrides in rare cases.

I have tweaked a couple of inconsistencies between the signup
content item and the finder itself ("Country" vs "Countries", for
example), but nothing that will change behaviour (topic names are
unchanged).

The biggest change is in the property tribunal decisions finder.
It seems some additional categories were added in #2360, but were
forgotten to be added to the corresponding email signup page.

I've also added unit tests for all the new functionality, which
required exposing the `email_filter_facets` method publicly and
also making it a static method to avoid having to initialise the
`FinderSignupContentItemPresenter` with a schema. I'm not entirely
pleased with the resulting code, which could do with a refactor.

Finally, note that I've removed the TODO comment around removing
'email_filter_by' when it's removed from finder-frontend. The
comment was added in December 2019 in be8462d0551b3997fdb1c8ad48cf9ee0cec9053a
and the field hasn't been removed yet, nor is it on anyone's
radar as far as I can tell, so leaving the comment would just be
misleading the reader.