Remove Email Alert Configuration spec and move test
This spec file was too tightly coupled to how email alerts are
configured in the finder schema itself, and not in testing that
the presented email signup page document type itself is well
formed.

There was one relevant test, about subject length, which I've
ported over to the more appropriate file. We can now delete the
old redundant test file.