Change form to allow user to choose 'email_filter_by' value
Now that we no longer need to support hash prefixes, we can more
easily expose the "email_filter_by" value to the user. It's also
not that difficult to let the user know all of the available
options for that field - why make lives more difficult for
ourselves by adding a "Request assistance from the developers"
kind of checkbox to the form?

So this is a tweak to the original design: we'll now expose all of
the possible 'email_filter_by' options in a dropdown, allowing the
user to better self-serve. To support this, we have a new (transient)
`email_filter_by_candidates` method on EmailAlert, to quickly
retrieve the available values for that property.