Remove TODO comment around checking permissions
GDS or departmental 'editors', or those who have been given
[‘editor’ permission for a specific finder](https://docs.publishing.service.gov.uk/repos/specialist-publisher/creating-editing-and-removing-specialist-document-types-and-finders.html#7-permissions),
can use the new "request edits" forms. This is the same subset
of people that can publish documents. The only other role
Specialist Publisher has a concept of is the
[writer role](https://github.com/alphagov/specialist-publisher/blob/5c0cb882d5e27f0e8979d1433a861491e160e486/app/policies/application_policy.rb#L18-L20),
which just checks that the user's organisation owns the document
type. Folks in that role are restricted to viewing and editing
content, but can't publish/unpublish.

Requiring a role of 'editor' seems a sensible precaution.
Double-checked in this Slack thread:
https://gds.slack.com/archives/C03D4A72HGE/p1737110535605969

We don't need to introduce an additional layer of permissions for
these forms.