Fix issue where email_filter_options is set as "null" string
Where there were no `email_filter_options` set (e.g. on AAIB
reports finder), the corresponding input was defaulting to a value
of "null" due to us rendering `to_json` directly into the value.
This then meant the parameter wasn't missing (or blank), and thus
the controller would try to `JSON.parse` the parameter, and break.

So we now:

1. Don't set a value on the hidden input, if `email_filter_options`
   is `nil`.
2. If the submitted value is an empty string, we account for that.
   (In slightly different ways depending on if we're in "all content"
   or "filtered content" mode, since "all content" would require
   no `email_filter_options` to be generated, whereas "filtered
   content" would require an `email_filter_options` hash to be
   generated so that we can set `email_filter_options.email_filter_by`.