Use JSON.pretty_generate on the proposed schema, for minimal diff
The [test ticket](https://govuk.zendesk.com/agent/tickets/6015210)
that came through brought in minified JSON, meaning if we were to
copy and paste the JSON into GitHub, we'd have a large diff.
Currently, developers would have to put the JSON through some
pre-processing before opening the PR.

Making this change breaks the current "POST zendesk" test, but
I don't have a lot of confidence in the test as it stands. Notice
that we've now changed `CmaCase.finder_schema.as_json` to
`CmaCase.finder_schema.to_json` - this was because the former
wouldn't really render as JSON (we'd have things like
`"target_stack" => "live"` instead of using colons), so clearly
the current test wasn't catching all issues. The test relies
heavily on stubbing.

I therefore don't feel _too_ bad about commenting out aspects of
the test, to come back to later. Zendesk ticket creation is
something we can't reliably test locally (via unit tests or
through the local dev UI) at the moment, so rather than continue
to hack around with the stubbing, I'll rely on manually testing
on our remote test environments, with a note to fix these up
later where possible.

As an aside, I did spend a couple of hours trying to get a
working test here, but I kept facing the issue of the
`hash_including` failing to match even with deleting properties,
and the awkwardness of trying to extract out the JSON from the
string that has the "Developer - raise a PR..." prefix, and then
the awkwardness of trying to un-escape the numerous \\\\
backslashes that are present in the JSON in order to do a `eq`
comparison with `JSON.pretty_generate(CmaCase.finder_schema.to_json)`.
It's fiddlier than it looks, and in the meantime, the form just
isn't that useful to us if it's not outputting prettified JSON.